pyproject.toml
[tool.poetry]
name = "agent-carro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-adk = {extras = ["eval"], version = "^1.0.0"}
pydantic = "^2.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
